// src/__generated__/chargers.ts
/**
 * !!! This file is autogenerated do not edit by hand !!!
 *
 * Generated by: @databases/pg-schema-print-types
 * Checksum: Vxp80/n21Pz4XAum98ZqzbceqtzgeSSaolwqi5lWaxOyeD51/KTk7DR7tIRqiwm3S5Xh6rmZX26L/Is4LhjfmA==
 */

/* eslint-disable */
// tslint:disable

import {StatusEnum} from './_enums'
import Locations from './locations'

interface Chargers {
  /**
   * @default uuid_generate_v4()
   */
  charger_id: string & {readonly __brand?: 'chargers_charger_id'}
  location_id: Locations['location_id']
  status: StatusEnum
}
export default Chargers;

interface Chargers_InsertParameters {
  /**
   * @default uuid_generate_v4()
   */
  charger_id?: string & {readonly __brand?: 'chargers_charger_id'}
  location_id: Locations['location_id']
  status: StatusEnum
}
export type {Chargers_InsertParameters}